add Test Methods With correct results

+++ src/main/java/pro/sky/java/course2/h_m_2_11_mockito/controller/DepartmentController.java
@@ -27,7 +27,7 @@ public class DepartmentController {
     }
 
     @GetMapping("{id}/salary/sum")
-    public int sumSalariesByDepartment(@PathVariable("id") int departmentId) {
+    public double sumSalariesByDepartment(@PathVariable("id") int departmentId) {
         return departmentService.sumSalariesByDepartment(departmentId);
     }
