+++ src/main/java/pro/sky/java/course2/h_m_2_11_mockito/service/DepartmentService.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.Optional;
 
 public interface DepartmentService {
-    int sumSalariesByDepartment (int department);
+    double sumSalariesByDepartment (int department);
     Optional<Employee> maxSalaryForDepartment(int department);
 
     Optional<Employee> minSalaryForDepartment(int department);
